+++ src/state/Game.hpp
@@ -31,7 +31,7 @@ class Game : public State {
   explicit Game(StateEngine& engine) : State(engine) {}
 
   void init() override;
-  void update(const double deltaTime) override;
+  void update(const float deltaTime) override;
   void draw() override;
   void cleanup() override {
     // No cleanup
